+++ Hooks/useHandleToolBar.js
@@ -3,17 +3,10 @@ import copyToClipboard from '../Services/copyToClipboard';
 
 import {useDispatch} from 'react-redux';
 import startPaste from '../Services/rnfs/startPaste';
-import askToRename from '../Services/askToRename';
-import moveItem from '../Services/rnfs/moveItem';
-import modalPromise from '../Actions/modalPromise';
-import Confirm from '../Layout/Modal/ModalBodies/Confirm';
-import collectHighilightedItems from '../Services/collectHighilightedItems';
-import MaterialIcon from '../Common/MaterialIcon/MaterialIcon';
-import {Text, View} from 'react-native';
-import styles from '../styles/styles';
-import useIcon from './useIcon';
-import SmallGrayText from '../Common/SmallGrayText/SmallGrayText';
-import DeleteProgress from '../Layout/Modal/ModalBodies/DeleteProgress';
+import handleDelete from '../Services/fileUtils/handleDelete';
+import handleRename from '../Services/fileUtils/handleRename';
+import handleNewFile from '../Services/fileUtils/handleNewFile';
+import handleNewFolder from '../Services/fileUtils/handleNewFolder';
 
 export default function useHandleToolBar(
   option,
@@ -35,83 +28,23 @@ export default function useHandleToolBar(
         break;
       }
       case 'paste': {
-        startPaste(dispatch, state.clipboardItems, {...tab});
+        startPaste(dispatch, state.clipboardItems, tab);
         break;
       }
       case 'delete': {
-        async function asyncDelete() {
-          const highlightedItems = collectHighilightedItems(filesList);
-          const isConfirmDelete = await modalPromise(
-            dispatch,
-            Confirm,
-            {
-              description: (
-                <View style={[styles.bigGap]}>
-                  <Text style={[styles.text]}>
-                    Do you want to delete these {highlightedItems.length} items?
-                  </Text>
-                  {highlightedItems.map(item => (
-                    <View
-                      key={item.path}
-                      style={[styles.rowLayout, styles.mediumGap]}>
-                      {useIcon(item)}
-                      <View style={[styles.wide]}>
-                        <Text
-                          ellipsizeMode="tail"
-                          numberOfLines={1}
-                          style={[styles.text]}>
-                          {item.name}
-                        </Text>
-                        <SmallGrayText ellipsizeMode="tail" numberOfLines={2}>
-                          {item.path}
-                        </SmallGrayText>
-                      </View>
-                    </View>
-                  ))}
-                </View>
-              ),
-            },
-            {
-              icon: <MaterialIcon name="delete-outline" />,
-              heading: 'Confirm Delete?',
-            },
-          );
-          if (isConfirmDelete)
-            await modalPromise(
-              dispatch,
-              DeleteProgress,
-              {
-                items: highlightedItems,
-              },
-              {
-                icon: <MaterialIcon name="progress-clock" />,
-                heading: `Delete in progress...`,
-                isStatic: true,
-              },
-            );
-          dispatch({type: 'TOAST', payload: 'Items deleted.'});
-          dispatch({type: 'SETREFRESHPATH', payload: tab.path});
-          dispatch({type: 'POPMODALSTACK'});
-        }
-        asyncDelete();
+        handleDelete(dispatch, filesList, tab);
         break;
       }
       case 'rename': {
-        async function asyncRename() {
-          let lastHighlightedItem = {
-            ...filesList.findLast(item => item.isHighlighted),
-          };
-          const {path, parent} = lastHighlightedItem;
-          lastHighlightedItem.destFilePath = parent;
-          const newName = await askToRename(dispatch, lastHighlightedItem);
-          if (newName) {
-            await moveItem(path, parent, newName);
-            dispatch({type: 'TOAST', payload: 'Renamed successfully.'});
-          }
-          dispatch({type: 'SETREFRESHPATH', payload: tab.path});
-          dispatch({type: 'POPMODALSTACK'});
-        }
-        asyncRename();
+        handleRename(dispatch, filesList, tab);
+        break;
+      }
+      case 'newFile': {
+        handleNewFile(dispatch, tab);
+        break;
+      }
+      case 'newFolder': {
+        handleNewFolder(dispatch, tab);
         break;
       }
     }
