+++ Services/Rnfs/newFile.js
@@ -1,0 +1,4 @@
+import RNFS from 'react-native-fs';
+export default async function newFile(destPath, name) {
+  return await RNFS.writeFile(destPath + '/' + name, '', 'utf8');
+}
